// src/app/models/preset.ts
export interface Preset {
    id: number;
    name: string;
    company: string;
    mattype: string;
    materialId: number;
    materialWidth: number;
    autoWidth: boolean;
    finishingId: number;
    quantity: number;
    height: number;
    width: number;
    scrap: string;
    manualScrap: number;
    description: string;
}